Split out image_builder code into its own package
Issue #479
